tooling/pyproject.toml
[project]
name = "doc-sync"
version = "0.1.0"
description = "Claude-powered API documentation sync tooling for any Mintlify docs site"
requires-python = ">=3.11"
dependencies = [
    "anthropic[bedrock]>=0.40.0",
    "python-frontmatter>=1.1.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
doc-sync      = "doc_sync.cli.sync:main"
doc-inventory = "doc_sync.cli.inventory:main"
doc-coverage  = "doc_sync.cli.coverage:main"
doc-stub      = "doc_sync.cli.stub:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["doc_sync"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
]
